qsort.h: first take, implemented partitioning
This incorporates some of my findings from 2013.

Producing a general-purpose sort routine can be hard, because the cost
of comparison is unknown.  Quicksort works best when comparisons are
cheap (e.g. comparing two integers takes roughly one instruction, while
swapping two integers takes roughly three instructions).  When
comparisons are expensive (e.g. strcmp), minimizing the number of
comparisons becomes very important; hence merge sort might work best
in this case.

Quicksort with median-of-3 and insertion sort cut-off makes about
1.2n log_2 n - 3n comparisons on the average.  Merge sort makes
n log_2 n comparisons in the worst case (and close to that in the
average case), roughly 10% fewer.  To some extent, this justifies the
use of merge sort in glibc.  Glibc's merge sort is somewhat suboptimal,
though: it has extra branches in the inner loop (instead of choosing the
right loop), and it calls memcpy to swap elements (instead of inlining
and/or predefining a bunch of copyN functions).  But optimizing glibc's
qsort is another story.

The upshot is that both rdtsc cycles AND the number of comparisons
must be taken into account.

For example, consider this partitioning loop:

    while (1) {
	do q_i++; while (Q_LESS(q_i, q_l));
	do q_j--; while (Q_LESS(q_l, q_j));
	if (q_i >= q_j) break;
	Q_SWAP(q_i, q_j);
    }

The loop is very tight, but it makes one extra useless iteration when
q_i + 1 == q_j.  There's then the temptation to amend the loop like this:

	Q_SWAP(q_i, q_j);
+	if (q_i + 1 == q_j) {
+	    q_i++, q_j--;
+	    break;
+	}

It is interesting to estimate both theoretically and practically what
kind of gains can such an optimization bring.

Let's assume that the optimization can save 1 comparison per partition.
Perfect but naiive Quicksort make 1 + 2 + 4 + ... + n/2, or about n,
partitions.  Thus the optimization can save n / (n log_2 n), or about
10%, of comparisons (these are very rough estimates indeed).
A non-naiive Quicksort, on the other hand, the one which switches to
insertion sort after the size of partition becomes 16 or less,
makes much fewer partitions.  Specifically, for each partition of
size 16, it skips 14 subsequent partitions:

	   8       8
	 4   4   4   4
	2 2 2 2 2 2 2 2

So, with the insertion-sort cut-off, the reduction in comparisons should
be below 1%.  Indeed, test runs show that with n=1024, about 0.3% to 0.4%
fewer comparisons are made.  (With these test runs, though, there's
another problem: enough of them must be done with srand(getpid()),
due to the dependency on input data inherent to Quicksort.)  And so
I conclude that it is probably not worth to amend the loop.