use (l+1,mid,r) median instead of (l,mid,r)
Using the median of (l,mid,r) leads to bad performance on decreasingly
sorted arrays.  I've just read about it in "BlockQuicksort: How Branch
Mispredictions don't affect Quicksort" by Stefan Edelkamp and Armin
Weiß, 2016.  The problem at first seemed specific to gcc's std::sort,
which uses somewhat different partitioning (for one, the median is not
put in its final place after partitioning - rather, the recursive
process is applied to the adjacent subfiles).  So I was surprised
to be able to reproduce this kind of bad behaviour with Q_PARTITION.

So let's see how we get there.
We start by comparing three elements:

	9 8 7 6 5 4 3 2 1 0
	^       ^         ^

and putting the median at the beginning:

	5 8 7 6 0 4 3 2 1 9
	^       ^         ^

The smallest element then goes to the second position:

	5 0 7 6 8 4 3 2 1 9
	  ^     ^

after which the loop starts:

	5 0 7 6 8 4 3 2 1 9
	    `-->     <--'

The loop exchanges the elements and finds out the right place for the
pivot:

	5 0 1 2 3 4 8 6 7 9
	          ^

The median is put in its right place, so finally we have the following
arrangement:

	4 0 1 2 3 5 8 6 7 9
	^       ^ ^ ^     ^
	m-1   m-2 m max-1 max

So, the problem is that, with the median being m, the recursive process
will use the median of (m-1,m-2,x)=m-2 for the left subfile, and the
median of (max-1,max,y)=max-1 for the right subfile.  Which are indeed
very bad choices.  It takes some effort to prove that these bad choices
recursively proliferate themselves.

Now if we take (l+1,mid,r) for the median, we end up with a similar
arrangement:

	4 0 1 2 3 5 9 6 7 8
	  ^     ^ ^   ^   ^
	          m

The elements differ only in the right subfile; but the crux of the
matter is that the choice of the second element instead of the first
immediately disrupts the pattern for the subfiles: note that the
two elements for the median are no longer consecutive.

I wonder how the analysis could be done for the more general case,
without the sophomore's habit of resorting to 0-9 numbers and the like.
Perhaps some computer scientists have the powerful abstract thinking
which makes them envision truths, and which alas I mostly lack.
If you now how to develop such thinking quickly, drop me a line.